==> predict-hapiness/notebooks/01_WordCount_study.ipynb <==
# Load Libraries
import numpy as np
import pandas as pd
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.naive_bayes import GaussianNB
from sklearn.preprocessing import LabelEncoder
import re
from sklearn.feature_extraction.text import CountVectorizer, TfidfVectorizer
from sklearn.model_selection import cross_val_score
from sklearn.metrics import accuracy_score, make_scorer

# load data
train = pd.read_csv("../input/train.csv")

train.head()

## join data
alldata = train.copy()
alldata['Is_Response'] = [1 if x == 'happy' else 0 for x in alldata['Is_Response']]

### set target variable
train_feats = train.copy()
train_feats['Is_Response'] = [1 if x == 'happy' else 0 for x in train_feats['Is_Response']]
target = train_feats['Is_Response']

from sklearn import preprocessing

def countWord(text, word):
    return text.lower().count(word)

def countWordRatio(text, word):
    return text.lower().count(word) / len(text)

happy_df = train[train['Is_Response'] == 'happy']
not_happy_df = train[train['Is_Response'] == 'not happy']

from collections import Counter
from itertools import chain
import re

authors_notflat = [re.sub(r'[^A-Za-z0-9\s]',r'',a.lower()).split() for a in happy_df['Description']]
happy_counter = Counter(chain.from_iterable(authors_notflat))

authors_notflat = [re.sub(r'[^A-Za-z0-9\s]',r'',a.lower()).split() for a in not_happy_df['Description']]
not_happy_counter = Counter(chain.from_iterable(authors_notflat))

def get_keys(counter, num):
    happy_elements = []

    for a in counter.most_common(num):
        happy_elements.append(a[0])

    return happy_elements

happy_counter.most_common(10)

words_set = set(get_keys(happy_counter, 3000))|set(get_keys(not_happy_counter, 2500))

happy_counter.most_common(20)

words_set.remove('the')
words_set.remove('and')
words_set.remove('a')
words_set.remove('to')
words_set.remove('was')
words_set.remove('in')
words_set.remove('of')
words_set.remove('is')
words_set.remove('for')
words_set.remove('it')
words_set.remove('at')
words_set.remove('with')
words_set.remove('this')

len(words_set)

import numpy as np

np.save('common-words.npy', list(words_set))
