==> predict-hapiness/notebooks/02_cnn_embeddings_keras_836.ipynb <==
import numpy as np
np.random.seed(31)
import pandas as pd
from sklearn.cross_validation import train_test_split
from subprocess import check_output

#LOAD DATA

train = pd.read_csv("../input/train.csv")
test = pd.read_csv("../input/test.csv")
sample = pd.read_csv("../input/sample_submission.csv")
sample.head(2)

#CREATE TARGET VARIABLE
train["happy"] = (train.Is_Response=="happy")*1
train["not_happy"] = (train.Is_Response=="not happy")*1
train.drop("Is_Response", 1, inplace=True)
target_vars = ["happy", "not_happy"]
train.head(2)

#STEMMING WORDS
import nltk.stem as stm
import re
stemmer = stm.SnowballStemmer("english")
train["stem_text"] = train.Description.apply(lambda x: (" ").join([stemmer.stem(z) for z in re.sub("[^a-zA-Z0-9]"," ", x).split(" ")]))
test["stem_text"] = test.Description.apply(lambda x: (" ").join([stemmer.stem(z) for z in re.sub("[^a-zA-Z0-9]"," ", x).split(" ")]))
train.head(3)

#PROCESS TEXT: RAW
from keras.preprocessing.text import Tokenizer
tok_raw = Tokenizer()
tok_raw.fit_on_texts(train.Description.str.lower())
tok_stem = Tokenizer()
tok_stem.fit_on_texts(train.stem_text)
train["seq_text_stem"] = tok_stem.texts_to_sequences(train.stem_text)
test["seq_text_stem"] = tok_stem.texts_to_sequences(test.stem_text)
train.head(3)

#EXTRACT DATA FOR KERAS MODEL
from keras.preprocessing.sequence import pad_sequences
def get_keras_data(dataset, maxlen=20):
    X = {
        "stem_input": pad_sequences(dataset.seq_text_stem, maxlen=maxlen)
    }
    return X


maxlen = 60
dtrain, dvalid = train_test_split(train, random_state=123, train_size=0.85)
X_train = get_keras_data(dtrain, maxlen)
y_train = np.array(dtrain[target_vars])
X_valid = get_keras_data(dvalid, maxlen)
y_valid = np.array(dvalid[target_vars])
X_test = get_keras_data(test, maxlen)

n_stem_seq = np.max( [np.max(X_valid["stem_input"]), np.max(X_train["stem_input"])])+1

#KERAS MODEL DEFINITION
from keras.layers import Dense, Dropout, Embedding
from keras.layers import Flatten, Input, SpatialDropout1D
from keras.models import Model
from keras.optimizers import Adam 

def get_model():
    embed_dim = 50
    dropout_rate = 0.9
    emb_dropout_rate = 0.9
   
    input_text = Input(shape=[maxlen], name="stem_input")
    
    emb_lstm = SpatialDropout1D(emb_dropout_rate) (Embedding(n_stem_seq, embed_dim
                                                ,input_length = maxlen
                                                               ) (input_text))
    dense = Dropout(dropout_rate) (Dense(512) (Flatten() (emb_lstm)))
    
    output = Dense(2, activation="softmax")(dense)

    model = Model([input_text], output)

    model.compile(loss="categorical_crossentropy", optimizer='nadam', metrics=["accuracy"])
    return model

model = get_model()
model.summary()
    

#TRAIN KERAS MODEL
model = get_model()
model.fit(X_train, y_train, epochs=27
          , validation_data=[X_valid, y_valid]
         , batch_size=1024)

#MODEL EVALUATION
from sklearn.metrics import accuracy_score

preds_train = model.predict(X_train)
preds_valid = model.predict(X_valid)

print(accuracy_score(preds_train.argmax(axis=1), y_train[0:,1]))
print(accuracy_score(preds_valid.argmax(axis=1), y_valid[0:,1]))

# 0.924573915146
# 0.83647260274

#PREDICTION
preds = pd.DataFrame(model.predict(get_keras_data(train, maxlen)), columns=['cnn_target_1', 'cnn_target_2'])
submission = pd.concat([train["User_ID"],preds], 1)
submission.to_csv("./cnn_train.csv", index=False)
submission.head()

#PREDICTION
preds = pd.DataFrame(model.predict(X_test), columns=['cnn_target_1', 'cnn_target_2'])
submission = pd.concat([test["User_ID"],preds], 1)
submission.to_csv("./cnn_test.csv", index=False)
submission.head()
