==> predict-hapiness/notebooks/03_catboost_features_893.ipynb <==
### Introduction

# Load Libraries
import numpy as np
import pandas as pd
from nltk.corpus import stopwords
from nltk.stem import PorterStemmer
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.preprocessing import LabelEncoder
import re
from sklearn.model_selection import cross_val_score
from sklearn.metrics import accuracy_score, make_scorer

# load data
train_df = pd.read_csv("../input/train.csv")
test_df = pd.read_csv("../input/test.csv")

cnn_train_df = pd.read_csv("cat_countvec_train.csv")
cnn_test_df = pd.read_csv("cat_countvec_test.csv")

cnn2_train_df = pd.read_csv("cnn_train.csv")
cnn2_test_df = pd.read_csv("cnn_test.csv")

train_df = pd.merge(train_df, cnn_train_df, on='User_ID', how='left')
test_df = pd.merge(test_df, cnn_test_df, on='User_ID', how='left')

train_df = pd.merge(train_df, cnn2_train_df, on='User_ID', how='left')
test_df = pd.merge(test_df, cnn2_test_df, on='User_ID', how='left')

train_df.head()

stopwords_set = set(stopwords.words("english"))

## Number of words in the text ##
train_df["num_words"] = train_df["Description"].apply(lambda x: len(str(x).split()))
test_df["num_words"] = test_df["Description"].apply(lambda x: len(str(x).split()))

## Number of unique words in the text ##
train_df["num_unique_words"] = train_df["Description"].apply(lambda x: len(set(str(x).split())))
test_df["num_unique_words"] = test_df["Description"].apply(lambda x: len(set(str(x).split())))

## Number of characters in the text ##
train_df["num_chars"] = train_df["Description"].apply(lambda x: len(str(x)))
test_df["num_chars"] = test_df["Description"].apply(lambda x: len(str(x)))

## Number of stopwords in the text ##
train_df["num_stopwords"] = train_df["Description"].apply(lambda x: len([w for w in str(x).lower().split() if w in stopwords_set]))
test_df["num_stopwords"] = test_df["Description"].apply(lambda x: len([w for w in str(x).lower().split() if w in stopwords_set]))

import string

train_df["num_punctuations_space"] = train_df['Description'].apply(lambda x: x.count(' ') )
test_df["num_punctuations_space"] = test_df['Description'].apply(lambda x: x.count(' ') )

for c in ['I', 'f']:
    train_df["num_char_{}".format(c)] = train_df['Description'].apply(lambda x: x.count(c) ) / train_df["num_chars"]
    test_df["num_char_{}".format(c)] = test_df['Description'].apply(lambda x: x.count(c) ) / test_df["num_chars"]

for c in ['!']:
    train_df["num_char_{}".format(c)] = train_df['Description'].apply(lambda x: x.count(c) ) / train_df["num_chars"]
    test_df["num_char_{}".format(c)] = test_df['Description'].apply(lambda x: x.count(c) ) / test_df["num_chars"]

train_df["num_stopwords_factor"] = train_df["num_stopwords"] / train_df["num_words"]
test_df["num_stopwords_factor"] = test_df["num_unique_words"] / test_df["num_words"]

train_df.tail()

## join data
test_df['Is_Response'] = np.nan
alldata = pd.concat([train_df, test_df]).reset_index(drop=True)

# split the merged data file into train and test respectively
train_feats = alldata[~pd.isnull(alldata.Is_Response)]
test_feats = alldata[pd.isnull(alldata.Is_Response)]

### set target variable
from sklearn import preprocessing
le = preprocessing.LabelEncoder()
train_df['Is_Response_idx'] = le.fit_transform(train_df['Is_Response'])

le.classes_

features = train_df.columns
features = features.drop(['User_ID', 'Description', 'Browser_Used', 'Device_Used', 'Is_Response', 'Is_Response_idx'])
features

from sklearn.cross_validation import train_test_split

train_X, valid_X, train_y, valid_y = train_test_split(train_df[features].as_matrix(), 
                                                      train_df['Is_Response_idx'].as_matrix(),
                                                      test_size=0.2,
                                                      random_state=42)

print(train_X.shape)
print(valid_X.shape)                                           
print(train_y.shape)
print(valid_y.shape)

### CatBoost

Catboost is a new package recently launched by Yandex. It is said that it works well when the data has many categorical features. We'll use it on count data and see it our model improves.

## import library
from catboost import CatBoostClassifier,cv, Pool

## define the classifer model
model = CatBoostClassifier(
    learning_rate=0.03,
    rsm = 0.8,
    depth=8,
    iterations=1000,
    od_type = 'Iter',
    od_wait = 10,
    eval_metric='MultiClass',
    loss_function='MultiClass',
    verbose=True)

train_df['Is_Response_idx'].value_counts(normalize=True)

## train model
model.fit(train_X, train_y, eval_set=(valid_X, valid_y))

feat_importances = model.get_feature_importance(valid_X, valid_y)
feat_df = pd.DataFrame({'features':features, 'importances':feat_importances})
feat_df.sort_values(by='importances')

y_pred = model.predict(train_df[features])

from sklearn.metrics import accuracy_score
accuracy_score(train_df['Is_Response_idx'].as_matrix(), y_pred)
# 0.93139319839720536
# 0.93157299907531077
# 0.95204459056817015

## make predictions
preds = model.predict(test_df[features])
preds.flatten()

feat_df = pd.DataFrame()
feat_df['preds'] = preds.flatten()
feat_df['preds'].value_counts(normalize=True)

test_predict = model.predict(test_df[features])

## make submission
sub5 = pd.DataFrame()
sub5['User_ID'] = test_df['User_ID']
sub5['Is_Response'] = test_predict
sub5['Is_Response'] = ['happy' if x < 0.499 else 'not_happy' for x in sub5['Is_Response']]

sub5

sub5.to_csv('submission.csv', index=False)
